chore: Moved utility function from Deploy to utils
