+++ src/utils.py
@@ -54,3 +54,18 @@ def wait_until_ready(model_name, model_version):
         if status == ModelVersionStatus.READY:
             break
         time.sleep(1)
+
+def wait_for_deployment(model_name, model_version, stage='Staging'):
+    client = MlflowClient()
+    status = False
+    while not status:
+        model_version_details = dict(
+            client.get_model_version(name=model_name,version=model_version)
+            )
+        if model_version_details['current_stage'] == stage:
+            print(f'Transition completed to {stage}')
+            status = True
+            break
+        else:
+            time.sleep(2)
+    return status
